# backend/bin/Debug/net8.0/swagger.json
{
    "openapi": "3.0.1",
    "info": {
      "title": "Art Gallery",
      "description": "Backend service that provides resources for the beautiful art gallery.",
      "contact": {
        "name": "Tanay and Jeetasha",
        "email": "tanay4847.be22@chitkara.edu.in"
      }
    },
    "paths": {
      "/api/artfacts": {
        "get": {
          "tags": [
            "Artfact"
          ],
          "responses": {
            "200": {
              "description": "Success",
              "content": {
                "text/plain": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Artfact"
                    }
                  }
                },
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Artfact"
                    }
                  }
                },
                "text/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Artfact"
                    }
                  }
                }
              }
            }
          }
        },
        "post": {
          "tags": [
            "Artfact"
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Artfact"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Artfact"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/Artfact"
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Created",
              "content": {
                "text/plain": {
                  "schema": {
                    "$ref": "#/components/schemas/Artfact"
                  }
                },
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Artfact"
                  }
                },
                "text/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Artfact"
                  }
                }
              }
            }
          }
        }
      },
      "/api/artfacts/{id}": {
        "get": {
          "tags": [
            "Artfact"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "style": "simple",
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success",
              "content": {
                "text/plain": {
                  "schema": {
                    "$ref": "#/components/schemas/Artfact"
                  }
                },
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Artfact"
                  }
                },
                "text/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Artfact"
                  }
                }
              }
            },
            "404": {
              "description": "Not Found",
              "content": {
                "text/plain": {
                  "schema": {
                    "$ref": "#/components/schemas/ProblemDetails"
                  }
                },
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ProblemDetails"
                  }
                },
                "text/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ProblemDetails"
                  }
                }
              }
            }
          }
        },
        "put": {
          "tags": [
            "Artfact"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "style": "simple",
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Artfact"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Artfact"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/Artfact"
                }
              }
            }
          },
          "responses": {
            "204": {
              "description": "No Content"
            },
            "400": {
              "description": "Bad Request",
              "content": {
                "text/plain": {
                  "schema": {
                    "$ref": "#/components/schemas/ProblemDetails"
                  }
                },
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ProblemDetails"
                  }
                },
                "text/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ProblemDetails"
                  }
                }
              }
            },
            "404": {
              "description": "Not Found",
              "content": {
                "text/plain": {
                  "schema": {
                    "$ref": "#/components/schemas/ProblemDetails"
                  }
                },
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ProblemDetails"
                  }
                },
                "text/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ProblemDetails"
                  }
                }
              }
            }
          }
        },
        "delete": {
          "tags": [
            "Artfact"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "style": "simple",
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "responses": {
            "204": {
              "description": "No Content"
            },
            "404": {
              "description": "Not Found",
              "content": {
                "text/plain": {
                  "schema": {
                    "$ref": "#/components/schemas/ProblemDetails"
                  }
                },
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ProblemDetails"
                  }
                },
                "text/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ProblemDetails"
                  }
                }
              }
            }
          }
        }
      },
...
    },
    "components": {
      "schemas": {
...
        "Artfact": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "format": "int32"
            },
            "year": {
              "type": "integer",
              "format": "int32"
            },
            "image": {
              "type": "string",
              "nullable": true
            },
            "description": {
              "type": "string",
              "nullable": true
            }
          },
          "additionalProperties": false
        },
...
        "ProblemDetails": {
          "type": "object",
          "properties": {
            "type": {
              "type": "string",
              "nullable": true
            },
            "title": {
              "type": "string",
              "nullable": true
            },
            "status": {
              "type": "integer",
              "format": "int32",
              "nullable": true
            },
            "detail": {
              "type": "string",
              "nullable": true
            },
            "instance": {
              "type": "string",
              "nullable": true
            }
          },
          "additionalProperties": {
  
          }
        },
...
      }
    }
  }